feat: resolvido comentario PR
