+++ src/Enums/CompanyServices.ts
@@ -1,0 +1,6 @@
+export enum COMPANY_SERVICES {
+  MAINTENANCE_SCHEDULED,
+  TECHNICAL_RESPONSIBILITY,
+  MAINTENANCE,
+  INSTALLATION
+}
