+++ src/Enums/TechnicianServices.ts
@@ -2,10 +2,3 @@ export enum TECHNICAL_MANAGER_SERVICES {
   AUTONOMOUS,
   EMPLOYEE_OF_A_COMPANY
 }
-
-export enum COMPANY_SERVICES {
-  MAINTENANCE_SCHEDULED,
-  TECHNICAL_RESPONSIBILITY,
-  MAINTENANCE,
-  INSTALLATION
-}
